Split write and read buffering into two classes
